+++ src/text/text_style.rs
@@ -1,10 +1,11 @@
 use crate::text::{
     renderer::{CharacterStyle, TextRenderer},
-    Alignment, Baseline,
+    Alignment, Baseline, LineHeight,
 };
 
 /// Text style.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
+#[non_exhaustive]
 pub struct TextStyle<S> {
     /// Character style.
     pub character_style: S,
@@ -14,12 +15,13 @@ pub struct TextStyle<S> {
 
     /// Text baseline.
     pub baseline: Baseline,
+
+    /// Line height.
+    pub line_height: LineHeight,
 }
 
-impl<S> TextStyle<S> {}
-
 /// Text style builder.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
 pub struct TextStyleBuilder<S> {
     style: TextStyle<S>,
 }
@@ -32,6 +34,7 @@ impl TextStyleBuilder<UndefinedCharacterStyle> {
                 character_style: UndefinedCharacterStyle,
                 alignment: Alignment::Left,
                 baseline: Baseline::Alphabetic,
+                line_height: LineHeight::default(),
             },
         }
     }
@@ -52,6 +55,13 @@ impl<S> TextStyleBuilder<S> {
         self
     }
 
+    /// Sets the line height.
+    pub fn line_height(mut self, line_height: LineHeight) -> Self {
+        self.style.line_height = line_height;
+
+        self
+    }
+
     /// Sets the character style.
     pub fn character_style<Style>(self, character_style: Style) -> TextStyleBuilder<Style> {
         TextStyleBuilder {
@@ -59,6 +69,7 @@ impl<S> TextStyleBuilder<S> {
                 character_style,
                 alignment: self.style.alignment,
                 baseline: self.style.baseline,
+                line_height: self.style.line_height,
             },
         }
     }
@@ -86,10 +97,8 @@ mod tests {
         pixelcolor::BinaryColor,
     };
 
-    // TODO: add tests
-
     #[test]
-    fn builder_alignments() {
+    fn builder() {
         let character_style = MonoTextStyleBuilder::<BinaryColor, _>::new()
             .font(Font6x9)
             .build();
@@ -98,9 +107,29 @@ mod tests {
             .character_style(character_style)
             .alignment(Alignment::Right)
             .baseline(Baseline::Top)
+            .line_height(LineHeight::Pixels(123))
             .build();
 
+        assert_eq!(text_style.character_style, character_style);
         assert_eq!(text_style.alignment, Alignment::Right);
         assert_eq!(text_style.baseline, Baseline::Top);
+        assert_eq!(text_style.line_height, LineHeight::Pixels(123));
+    }
+
+    #[test]
+    fn builder_default() {
+        let character_style = MonoTextStyleBuilder::<BinaryColor, _>::new()
+            .font(Font6x9)
+            .build();
+
+        // Setting a `character_style` is required to be able to call `build`.
+        let text_style = TextStyleBuilder::new()
+            .character_style(character_style)
+            .build();
+
+        assert_eq!(text_style.character_style, character_style);
+        assert_eq!(text_style.alignment, Alignment::Left);
+        assert_eq!(text_style.baseline, Baseline::Alphabetic);
+        assert_eq!(text_style.line_height, LineHeight::Percent(100));
     }
 }